# lib/api-zod/src/generated/types/experience.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { I18nList } from "./i18nList";
import type { I18nText } from "./i18nText";

export interface Experience {
  /** Stable semantic ID (e.g. 'yacht', 'atv') */
  id: string;
  sortOrder: number;
  visible: boolean;
  tagType: string;
  category: string[];
  imageUrl: string;
  fallbackEmoji: string;
  price: number;
  durationHours: string;
  title: I18nText;
  desc: I18nText;
  includes: I18nList;
  updatedAt: Date;
}