# lib/api-zod/src/generated/types/i18nList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Per-language string arrays. Keys are language codes.
 */
export interface I18nList {
  [key: string]: string[];
}