# lib/api-zod/src/generated/types/i18nText.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Per-language text values. Keys are language codes (en, es, it, fr, ...).
 */
export interface I18nText {
  [key: string]: string;
}