# lib/api-zod/src/generated/types/updateExperienceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { I18nText } from "./i18nText";

/**
 * Fields to update on an experience. All fields are optional.
 */
export interface UpdateExperienceBody {
  visible?: boolean;
  imageUrl?: string;
  sortOrder?: number;
  title?: I18nText;
  desc?: I18nText;
}